Feat: RecordCreate.vue -> 기록제목 날짜 내용과 이미지 업로드 부분 features/ui 로 기능 분리
